backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
email-validator==2.1.0
python-dotenv==1.0.0
boto3==1.29.0
redis==5.0.1
pytest==7.4.3
httpx==0.25.2
